pyproject.toml
[tool.poetry]
name = "baseer"
version = "0.0.1"
description = "BASEER AI - Arabic Image Captioning with Arabic culture object and landmark detection"
authors = ["Abdulmalik <reex4des@gmail.com>", "Saud <cs.saud.alashrah@gmail.com>"]
readme = "README.md"
packages = [{include = "baseer"}]


[tool.poetry.dependencies]
python = "^3.8"
matplotlib = "^3.7.1"
ftfy = "^6.1.1"
regex = "^2023.5.5"
tqdm = "^4.65.0"
textblob = "^0.17.1"
nltk = "^3.8.1"
spacy = "^3.5.3"
translate = "^3.6.1"
ipyplot = "^1.1.1"
torch = "^1.7.1"
pillow = "^9.5.0"
clip = { git = "https://github.com/openai/CLIP.git" }
salesforce-lavis = "^1.0.2"
numpy = "^1.24.3"

[tool.poetry.dependencies.en_core_web_lg]
url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0.tar.gz"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.1"
ipywidgets = "^8.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

